refactor files, add console
